Apply suggestions from code review
gas saver

+++ contracts/modules/TransferManager/SignedTransferManager.sol
@@ -58,7 +58,7 @@ contract SignedTransferManager is ITransferManager {
     */
     function updateSigners(address[] _signers, bool[] _signersStats) public withPerm(ADMIN) {
         require(_signers.length == _signersStats.length, "input array length does not match");
-        for(uint8 i=0; i<_signers.length; i++){
+        for(uint256 i=0; i<_signers.length; i++){
             signers[_signers[i]] = _signersStats[i]; 
         }
         emit UpdateSigners(_signers, _signersStats);
